requirements.txt
Pillow
cryptography
bitarray
pyperclip
streamlit
